Organized all components into subfolders and added an index.js to export all components. Removed unused variables and styles.
