--- src/components/Button.js
@@ -1,25 +1,0 @@
-import styled from 'styled-components'
-
-import React from 'react'
-
-const Button = ({content}) => {
-  return (
-    <StyledButton>{content}</StyledButton>
-  )
-}
-
-const StyledButton = styled.button`
-
-background-color: #fffff;
-color: #292E36;
-border-color: #ffff;
-border: 1px;
-border-style: solid;
-width: 80%;
-height: 40px;
-text-transform: uppercase;
-cursor: pointer;
-
-`
-
-export default Button
