+++ src/components/Shared/SocialStack.js
@@ -1,9 +1,9 @@
 import React from "react";
 import styled from "styled-components";
-import fbLogo from "../assets/icons/fb.png";
-import instaLogo from "../assets/icons/insta.png";
-import pinterestLogo from "../assets/icons/pinterest.png";
-import twitterLogo from "../assets/icons/twitter.png";
+import fbLogo from "../../assets/icons/fb.png"
+import instaLogo from "../../assets/icons/insta.png";
+import pinterestLogo from "../../assets/icons/pinterest.png";
+import twitterLogo from "../../assets/icons/twitter.png";
 
 const SocialStackWrapper = styled.div`
   display: flex;
