+++ src/components/index.js
@@ -1,3 +1,9 @@
-import Services from "./Services";
+import Hero from "./Hero/Hero"
+import Services from "./Home/Services"
+import Footer from "./Shared/Footer"
+import Header from "./Shared/Header"
+import SocialStack from "./Shared/SocialStack"
+import Input from "./User/Input"
+import SignIn from "./User/SignIn"
 
-export { Services };
+export {Hero, Services, Footer, Header, SocialStack, Input, SignIn};
