--- src/styles/SignIn.css
@@ -1,17 +1,0 @@
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-  }
-  
-  body {
-    font-family: 'Roboto', sans-serif;
-    background-image: url('/public/images/bg-signin.jpg');
-    background-size: cover;
-    display: flex ;
-    justify-content: center;
-    align-items: center;
-    height: 100vh;
-    width: 100vw;
-  }
-
